% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Contiunuous_cor.R
\name{Continuous_cor}
\alias{Continuous_cor}
\title{Creating a correlation matrix for all numeric variables}
\usage{
Continuous_cor(year)
}
\arguments{
\item{The}{Year in which the user wants to find out}
}
\value{
The correlation matrix for all numeric variables
}
\description{
This function finds out the correlation matrix
for all numeric variables in the data given the user input year.
}
