% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Leading_scorer.R
\name{leading_scorer}
\alias{leading_scorer}
\title{Finding out highest scoring player}
\usage{
leading_scorer(year)
}
\arguments{
\item{The}{Year in which the user wants to find out}
}
\value{
The record of the highest scoring player in the given year
}
\description{
This function finds out the highest scoring player given the user input year.
}
\keyword{HW5nba}
