% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Finding out oldest player}
\usage{
oldest_player(year)
}
\arguments{
\item{The}{Year in which the user wants to find out}
}
\value{
The record of the oldest player in the given year
}
\description{
This function finds out the older player given the user input year.
in the case of player with the same age found,
The player on the top of the filter result is going to be returned
}
\keyword{HW5nba}
